data/pyproject.toml
[tool.poetry]
name = "swisshikefinder"
version = "0.0.1"
description = "transform data for swiss hike finder"
readme = "README.md"
authors = ["Jan Benisek <benisekj@gmailcom>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
httpx="^0.27.0"
ipython = "^8.20.0"
ipykernel = "^6.28.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"